系统结构优化
1. CameraModule和VisionUI两个模块，CameraModule专注于采集数据，VisionUI提供主界面，提供接口让其他模块使用;
2. VisionDetect再拆分，操作界面设置到主界面右边，VisionDetect的VisionView删除，直接调用主界面的VisionView显示
