+++ CarameModel/cameramodule.h
@@ -4,13 +4,12 @@
 #include "caramemodel_global.h"
 #include "../include/ICamera.h"
 #include "../common/modulebase.h"
-#include "QMainView.h"
-#include "CameraCtrl.h"
-#include "viewctrl.h"
 
 #include "opencv/cv.h"
 
+class CameraCtrl;
 class CameraSetting;
+class QMainProcess;
 class CameraModule : public QModuleBase, public ICamera
 {
 public:
@@ -36,12 +35,6 @@ public:
 	void unlockCameraCapture();
 	bool isCameraCaptureAvaiable();
 
-	void addImageText(QString& szText);
-	void clearImage();
-	void setImage(cv::Mat& image, bool bDisplay);
-
-	void setHeightData(cv::Mat& matHeight);
-
 public:
 	 void openCamera();
 	 void closeCamera();
@@ -49,9 +42,6 @@ public:
 
 	 bool grabCamImage(int nCamera, cv::Mat& image, bool bSync);
 	 void setExposureTime(int nCamera, double exposureTime);	
-
-	 //»ñÈ¡Í¼Ïñ´°¿Ú	
-	 virtual QWidget* getMainView();
 
 	 void load3DViewData(int nSizeX, int nSizeY, QVector<double>& xValues, QVector<double>& yValues, QVector<double>& zValues);
 	 void show3DView();
@@ -64,7 +54,7 @@ private:
 private: 
 	CameraCtrl*      m_pCameraCtrl;
 	CameraSetting* m_pSetting;
-	QMainView* m_pMainView;
+	QSharedPointer<QMainProcess> m_pMainProcess;
 };
 
 #endif // CARAMEMODEL_H
