+++ VisionFormatWork2.sln
@@ -68,6 +68,7 @@ Global
 		{1795F236-00DB-4229-A2E2-FD21754B6206}.Debug|Win32.ActiveCfg = Debug|Win32
 		{1795F236-00DB-4229-A2E2-FD21754B6206}.Debug|Win32.Build.0 = Debug|Win32
 		{1795F236-00DB-4229-A2E2-FD21754B6206}.Debug|x64.ActiveCfg = Debug|x64
+		{1795F236-00DB-4229-A2E2-FD21754B6206}.Debug|x64.Build.0 = Debug|x64
 		{1795F236-00DB-4229-A2E2-FD21754B6206}.Release|Win32.ActiveCfg = Release|Win32
 		{1795F236-00DB-4229-A2E2-FD21754B6206}.Release|Win32.Build.0 = Release|Win32
 		{1795F236-00DB-4229-A2E2-FD21754B6206}.Release|x64.ActiveCfg = Release|x64
